refactor import methods into separate classes for better readability
